refactor: clean task execution panel internals
